+++ src/getStepStartStates.ts
@@ -106,7 +106,7 @@ import {
   STEP_ACCESS_CONTEXT_MANAGER_ACCESS_POLICIES,
   STEP_ACCESS_CONTEXT_MANAGER_SERVICE_PERIMETERS,
 } from './steps/access-context-manager/constants';
-import { STEP_BILLING_BUDGET } from './steps/billing-budget/constants';
+import { STEP_BILLING_BUDGET } from './steps/cloud-billing/constants';
 
 function validateInvocationConfig(
   context: IntegrationExecutionContext<SerializedIntegrationConfig>,
