+++ src/index.ts
@@ -28,7 +28,7 @@ import { spannerSteps } from './steps/spanner';
 import { apiGatewaySteps } from './steps/api-gateway';
 import { privateCaSteps } from './steps/privateca';
 import { accessPoliciesSteps } from './steps/access-context-manager';
-import { billingBudgetSteps } from './steps/billing-budget';
+import { cloudBillingSteps } from './steps/cloud-billing';
 
 import { Client } from './google-cloud/client';
 import { cloudAssetSteps } from './steps/cloud-asset';
@@ -80,7 +80,7 @@ export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> =
       ...apiGatewaySteps,
       ...privateCaSteps,
       ...accessPoliciesSteps,
-      ...billingBudgetSteps,
+      ...cloudBillingSteps,
     ],
     dependencyGraphOrder: ['last'],
     beforeAddEntity(context, entity: Entity): Entity {
