--- src/steps/billing-budget/__recordings__/fetchBillingBudget_979499027/recording.har
@@ -1,253 +1,0 @@
-{
-  "log": {
-    "_recordingName": "fetchBillingBudget",
-    "creator": {
-      "comment": "persister:JupiterOneIntegationFSPersister",
-      "name": "Polly.JS",
-      "version": "4.3.0"
-    },
-    "entries": [
-      {
-        "_id": "acea721c8193b51ced888cae721cc423",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 739,
-          "cookies": [],
-          "headers": [
-            {
-              "_fromType": "array",
-              "name": "content-type",
-              "value": "application/x-www-form-urlencoded"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept",
-              "value": "application/json"
-            },
-            {
-              "_fromType": "array",
-              "name": "content-length",
-              "value": "739"
-            },
-            {
-              "_fromType": "array",
-              "name": "user-agent",
-              "value": "node-fetch/1.0 (+https://github.com/bitinn/node-fetch)"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept-encoding",
-              "value": "gzip,deflate"
-            },
-            {
-              "_fromType": "array",
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "host",
-              "value": "www.googleapis.com"
-            }
-          ],
-          "headersSize": 300,
-          "httpVersion": "HTTP/1.1",
-          "method": "POST",
-          "postData": {
-            "mimeType": "application/x-www-form-urlencoded",
-            "params": [],
-            "text": "[REDACTED]"
-          },
-          "queryString": [],
-          "url": "https://www.googleapis.com/oauth2/v4/token"
-        },
-        "response": {
-          "bodySize": 550,
-          "content": {
-            "mimeType": "application/json; charset=UTF-8",
-            "size": 550,
-            "text": "{\"access_token\":\"[REDACTED]\",\"expires_in\":9999,\"token_type\":\"Bearer\"}"
-          },
-          "cookies": [],
-          "headers": [
-            {
-              "name": "content-type",
-              "value": "application/json; charset=UTF-8"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, X-Origin, Referer"
-            },
-            {
-              "name": "date",
-              "value": "Thu, 29 Jul 2021 15:04:21 GMT"
-            },
-            {
-              "name": "server",
-              "value": "scaffolding on HTTPServer2"
-            },
-            {
-              "name": "cache-control",
-              "value": "private"
-            },
-            {
-              "name": "x-xss-protection",
-              "value": "0"
-            },
-            {
-              "name": "x-frame-options",
-              "value": "SAMEORIGIN"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "alt-svc",
-              "value": "h3=\":443\"; ma=2592000,h3-29=\":443\"; ma=2592000,h3-T051=\":443\"; ma=2592000,h3-Q050=\":443\"; ma=2592000,h3-Q046=\":443\"; ma=2592000,h3-Q043=\":443\"; ma=2592000,quic=\":443\"; ma=2592000; v=\"46,43\""
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            }
-          ],
-          "headersSize": 533,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2021-07-29T15:04:32.352Z",
-        "time": 136,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 136
-        }
-      },
-      {
-        "_id": "71bfec2220530909635dac04f4913b2c",
-        "_order": 0,
-        "cache": {},
-        "request": {
-          "bodySize": 0,
-          "cookies": [],
-          "headers": [
-            {
-              "_fromType": "array",
-              "name": "x-goog-api-client",
-              "value": "gdcl/5.0.2 gl-node/14.17.3 auth/7.1.1"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept-encoding",
-              "value": "gzip"
-            },
-            {
-              "_fromType": "array",
-              "name": "user-agent",
-              "value": "google-api-nodejs-client/5.0.2 (gzip)"
-            },
-            {
-              "_fromType": "array",
-              "name": "authorization",
-              "value": "[REDACTED]"
-            },
-            {
-              "_fromType": "array",
-              "name": "accept",
-              "value": "application/json"
-            },
-            {
-              "_fromType": "array",
-              "name": "connection",
-              "value": "close"
-            },
-            {
-              "name": "host",
-              "value": "billingbudgets.googleapis.com"
-            }
-          ],
-          "headersSize": 563,
-          "httpVersion": "HTTP/1.1",
-          "method": "GET",
-          "queryString": [],
-          "url": "https://billingbudgets.googleapis.com/v1/billingAccounts/01694E-12D60B-BEBCE5/budgets"
-        },
-        "response": {
-          "bodySize": 68,
-          "content": {
-            "mimeType": "application/json; charset=UTF-8",
-            "size": 68,
-            "text": "{}"
-          },
-          "cookies": [],
-          "headers": [
-            {
-              "name": "content-type",
-              "value": "application/json; charset=UTF-8"
-            },
-            {
-              "name": "vary",
-              "value": "Origin, X-Origin, Referer"
-            },
-            {
-              "name": "date",
-              "value": "Thu, 29 Jul 2021 15:04:22 GMT"
-            },
-            {
-              "name": "server",
-              "value": "ESF"
-            },
-            {
-              "name": "cache-control",
-              "value": "private"
-            },
-            {
-              "name": "x-xss-protection",
-              "value": "0"
-            },
-            {
-              "name": "x-frame-options",
-              "value": "SAMEORIGIN"
-            },
-            {
-              "name": "x-content-type-options",
-              "value": "nosniff"
-            },
-            {
-              "name": "alt-svc",
-              "value": "h3=\":443\"; ma=2592000,h3-29=\":443\"; ma=2592000,h3-T051=\":443\"; ma=2592000,h3-Q050=\":443\"; ma=2592000,h3-Q046=\":443\"; ma=2592000,h3-Q043=\":443\"; ma=2592000,quic=\":443\"; ma=2592000; v=\"46,43\""
-            },
-            {
-              "name": "connection",
-              "value": "close"
-            }
-          ],
-          "headersSize": 510,
-          "httpVersion": "HTTP/1.1",
-          "redirectURL": "",
-          "status": 200,
-          "statusText": "OK"
-        },
-        "startedDateTime": "2021-07-29T15:04:32.500Z",
-        "time": 1052,
-        "timings": {
-          "blocked": -1,
-          "connect": -1,
-          "dns": -1,
-          "receive": 0,
-          "send": 0,
-          "ssl": -1,
-          "wait": 1052
-        }
-      }
-    ],
-    "pages": [],
-    "version": "1.2"
-  }
-}
